% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_generics.R
\name{cdf}
\alias{cdf}
\title{Cumulative Distribution Function}
\usage{
cdf(x, ...)
}
\arguments{
\item{x}{An object representing a probability distribution.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A vector of probabilities corresponding to the calculated CDF.
}
\description{
A generic function to compute the Cumulative Distribution Function (CDF)
for a given distribution object.
}
