% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_class.R
\name{distrib-class}
\alias{distrib-class}
\alias{`distrib`}
\title{S3 Class for Probability Distributions}
\description{
A generic S3 object of class \code{"distrib"} that encapsulates all mathematical properties
of a probability distribution.
}
\details{
S3 Class \code{distrib}

Objects of class \code{"distrib"} are typically created by constructor functions such as
\code{\link{gaussian_distrib}}.
}
\section{Structure}{

An object of class \code{"distrib"} is a \code{list} with the following components:

\describe{
\item{\code{}}{

}
\item{\code{}}{

}
}
}

