% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Reciprocal (Inverse) Transformation}
\usage{
inverse_transform()
}
\value{
A list of class \code{"transformer"}.
}
\description{
Creates a transformer object for the reciprocal transformation \eqn{Y = 1/X}.
}
\details{
\strong{Transformation:} \eqn{Y = 1/X}

\strong{Inverse:} \eqn{X = 1/Y}

\strong{Jacobian:} \eqn{J = \dfrac{dX}{dY} = -\dfrac{1}{Y^2}} (Absolute value taken for density: \eqn{1/Y^2})

\strong{Log-Jacobian Derivatives:}
\itemize{
\item Gradient: \eqn{\dfrac{\partial \log|J|}{\partial y} = -\dfrac{2}{y}}
\item Hessian: \eqn{\dfrac{\partial^2 \log|J|}{\partial y^2} = \dfrac{2}{y^2}}
}

\strong{Inverse Hessian:} \eqn{\dfrac{d^2 X}{d Y^2} = \dfrac{2}{Y^3}}

\strong{Support:} \eqn{X \neq 0}. Typically used for \eqn{X > 0} or \eqn{X < 0}.
Since the function is strictly decreasing, bounds and quantiles are automatically swapped.
}
