% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_functions.R
\name{kurtosis.default}
\alias{kurtosis.default}
\title{Empirical Excess Kurtosis}
\usage{
\method{kurtosis}{default}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric scalar representing the excess kurtosis.
}
\description{
Calculates the coefficient of \strong{excess} kurtosis (fourth standardized moment minus 3) for a numeric vector.
}
\details{
Computes the sample excess kurtosis:
\deqn{g_2 = \frac{\frac{1}{n} \sum_{i=1}^n (x_i - \bar{x})^4}{ \left( \frac{1}{n} \sum_{i=1}^n (x_i - \bar{x})^2 \right)^2 } - 3}

A value of 0 implies the distribution has the same "tailedness" as a Normal distribution.
Positive values indicate heavier tails (leptokurtic), negative values indicate lighter tails (platykurtic).
}
