% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_generics.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate Kurtosis}
\usage{
kurtosis(x, ...)
}
\arguments{
\item{x}{An object representing a probability distribution.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A numeric vector representing the kurtosis coefficient.
}
\description{
A generic function to calculate the kurtosis (fourth standardized moment) of a
distribution object.
}
\details{
Kurtosis measures the "tailedness" of the probability distribution.
It is defined as:
\deqn{\beta_2 = E\left[\left(\dfrac{X - \mu}{\sigma}\right)^4\right] = \dfrac{\mu_4}{\sigma^4}}

\strong{Continuous Case:}
\deqn{\beta_2 = \dfrac{1}{\sigma^4} \int_{-\infty}^{\infty} (x - \mu)^4 f(x) \, dx}

\strong{Discrete Case:}
\deqn{\beta_2 = \dfrac{1}{\sigma^4} \sum_{x \in S} (x - \mu)^4 p(x)}

where \eqn{\mu_4} is the fourth central moment and \eqn{\sigma} is the standard deviation.
Note that this function typically returns the raw kurtosis (Normal distribution = 3).
Use \code{kurtosis - 3} to obtain the \emph{excess kurtosis}.
}
