% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_distrib.R
\name{negbin_distrib}
\alias{negbin_distrib}
\title{Negative Binomial \code{distrib} Object (NB2)}
\usage{
negbin_distrib(link_mu = log_link(), link_theta = log_link())
}
\arguments{
\item{link_mu}{A link function object for the mean parameter \eqn{\mu}.
Defaults to \code{\link[linkfunctions]{log_link}}.}

\item{link_theta}{A link function object for the dispersion parameter \eqn{\theta}.
Defaults to \code{\link[linkfunctions]{log_link}}.}
}
\value{
A list of class \code{"distrib"}.
}
\description{
Creates a distribution object for the Negative Binomial distribution (NB2)
parameterized by mean (\eqn{\mu}) and dispersion (\eqn{\theta}).
}
\details{
The PMF is: \deqn{P(Y=y; \mu, \theta) = \dfrac{\Gamma(y+\theta)}{y!\Gamma(\theta)} \left(\dfrac{\theta}{\theta+\mu}\right)^\theta \left(\dfrac{\mu}{\theta+\mu}\right)^y}

\strong{Moments:}
\itemize{
\item Expected value: \eqn{\mathbb{E}(y) = \mu}
\item Variance: \eqn{\mathbb{V}(y) = \mu + \dfrac{\mu^2}{\theta}}
\item Skewness: \eqn{\gamma_1 = \dfrac{\theta + 2\mu}{\sqrt{\mu\theta(\theta + \mu)}}}
\item Excess Kurtosis: \eqn{\gamma_2 = \dfrac{6}{\theta} + \dfrac{(\theta + \mu)^2}{\mu\theta(\theta + \mu)}}
}
}
