% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_method.R
\name{plot.distrib}
\alias{plot.distrib}
\title{Plot Method for \code{distrib} Objects}
\usage{
\method{plot}{distrib}(x, theta, xlim = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"distrib"}.}

\item{theta}{A named list or vector of parameters. Must match \code{x$params}.}

\item{xlim}{Optional numeric vector of length 2 indicating the x-axis range.
If \code{NULL} (default), the range is automatically calculated.}

\item{...}{Additional arguments passed to the base \code{\link{plot}} function.}
}
\description{
Visualizes the Probability Density Function (PDF) or Probability Mass Function (PMF)
of a distribution object for a specific set of parameters.
}
