% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_method.R
\name{print.distrib}
\alias{print.distrib}
\title{Print Method for \code{distrib} Objects}
\usage{
\method{print}{distrib}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"distrib"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The object \code{x} invisibly.
}
\description{
Print Method for \code{distrib} Objects
}
