% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_functions.R
\name{std_dev.distrib}
\alias{std_dev.distrib}
\title{Calculate Standard Deviation for Distribution Objects}
\usage{
\method{std_dev}{distrib}(x, theta, use_moment = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"distrib"}.}

\item{theta}{A named list of parameters.}

\item{use_moment}{Logical. If \code{TRUE}, forces numerical calculation via the
\code{\link{moment}} function, even if an analytical formula exists.}

\item{...}{Additional arguments passed to \code{\link{variance.distrib}}.}
}
\value{
A numeric vector representing the standard deviation.
}
\description{
The standard deviation is calculated as the square root of the variance.
See \code{\link{variance.distrib}} for details on the calculation strategy.
}
\details{
Computes the standard deviation of a distribution object.
}
