% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_generics.R
\name{std_dev}
\alias{std_dev}
\title{Calculate Standard Deviation}
\usage{
std_dev(x, ...)
}
\arguments{
\item{x}{An object representing a probability distribution.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A numeric vector representing the standard deviation.
}
\description{
A generic function to calculate the standard deviation of a distribution object.
}
\details{
The standard deviation is the square root of the variance:
\deqn{\sigma = \sqrt{Var(X)} = \sqrt{E[(X - \mu)^2]}}

It shares the same units as the random variable \eqn{X}.
}
\seealso{
\code{\link{variance}}
}
