% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_functions.R
\name{variance.distrib}
\alias{variance.distrib}
\title{Calculate Variance for Distribution Objects}
\usage{
\method{variance}{distrib}(x, theta, use_moment = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"distrib"}.}

\item{theta}{A named list of parameters. Vectors are supported.}

\item{use_moment}{Logical. If \code{TRUE}, forces numerical calculation via the
\code{\link{moment}} function, even if an analytical formula exists.}

\item{...}{Additional arguments passed to \code{\link{moment}}.}
}
\value{
A numeric vector representing the variance.
}
\description{
Calculates the variance analytically if the formula is available in the object.
If not (or if forced), it calculates the variance numerically.
}
\details{
Computes the variance (second central moment) of a distribution object.
}
\seealso{
\code{\link{std_dev.distrib}}
}
