% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_adjusted.R
\name{zero_adjusted}
\alias{zero_adjusted}
\title{Zero-Adjusted Distribution Wrapper}
\usage{
zero_adjusted(distrib, link_za = logit_link())
}
\arguments{
\item{distrib}{An object of class \code{"distrib"}.}

\item{link_za}{A link function object for the zero probability parameter.
Defaults to \code{\link[linkfunctions]{logit_link}}.}
}
\value{
A new object of class \code{"distrib"} representing the zero-adjusted version of the input distribution.
}
\description{
A generic constructor for creating Zero-Adjusted (ZA) distributions.
This function automatically detects the type of the input distribution (discrete or continuous)
and dispatches the call to the appropriate specialized constructor:
\itemize{
\item \strong{Discrete:} Calls \code{\link{zero_adjusted_discrete}} to create a Hurdle model.
\item \strong{Continuous:} Calls \code{\link{zero_adjusted_continuous}} to create a Zero-Inflated continuous model (ZAW).
}
}
\seealso{
\code{\link{zero_adjusted_discrete}}, \code{\link{zero_adjusted_continuous}}
}
