% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Reciprocal (Inverse) Transformation}
\usage{
inverse_transform()
}
\value{
A list of class \code{"transformer"}.
}
\description{
Creates a transformer object for the reciprocal transformation \eqn{Y = 1/X}.
}
\details{
\strong{Transformation:} \eqn{Y = 1/X}

\strong{Inverse:} \eqn{X = 1/Y}

\strong{Jacobian:} \eqn{J = \dfrac{dX}{dY} = -\dfrac{1}{Y^2}} (Absolute value taken for density: \eqn{1/Y^2})

\strong{Support:} \eqn{X \neq 0}. Typically used for \eqn{X > 0} or \eqn{X < 0}.
Since the function is strictly decreasing, bounds and quantiles are automatically swapped.
}
