% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_functions.R
\name{skewness.distrib}
\alias{skewness.distrib}
\title{Calculate Kurtosis for Distribution Objects}
\usage{
\method{skewness}{distrib}(x, theta, use_moment = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"distrib"}.}

\item{theta}{A named list of parameters.}

\item{use_moment}{Logical. If \code{TRUE}, forces numerical calculation via the
\code{\link{moment}} function, even if an analytical formula exists.}

\item{...}{Additional arguments passed to \code{\link{moment}}.}
}
\value{
A numeric vector representing the kurtosis.
}
\description{
Measures the "tailedness" of the distribution. It uses the analytical formula if available.
Otherwise, it computes the fourth central moment and divides it by the fourth power of the
standard deviation.
}
\details{
Computes the kurtosis (fourth standardized moment) of a distribution object.
}
