% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{yj_transform}
\alias{yj_transform}
\title{Yeo-Johnson Transformation}
\usage{
yj_transform(lambda)
}
\arguments{
\item{lambda}{Numeric. The transformation parameter.}
}
\value{
A list of class \code{"transformer"}.
}
\description{
Creates a transformer object for the Yeo-Johnson transformation, which extends Box-Cox to negative values.
}
\details{
\strong{Transformation:} Defined piecewise:
\itemize{
\item \eqn{X \ge 0}: Box-Cox of \eqn{X+1} with parameter \eqn{\lambda}.
\item \eqn{X < 0}: Box-Cox of \eqn{|X|+1} with parameter \eqn{2-\lambda}, negated.
}

\strong{Support:} Defined for all \eqn{X \in \mathbb{R}}.
}
